ResourceManagement

Documentation for ResourceManagement.

Base.:+

Base.:+

ResourceManagement.ReadAvailHours

ResourceManagement.ReadLaborTracker

ResourceManagement._TeamBuilder

ResourceManagement._TeamBuilder

ResourceManagement._getAvailMonthHours

ResourceManagement._getEmployeePlannedHours

ResourceManagement._pad_zerosEnd

ResourceManagement.fetchAndWritePlannedHours!

ResourceManagement.getCapacity

ResourceManagement.getFwdPlannedHours

ResourceManagement.getUtilization

ResourceManagement.writeAvailableFwdHours!

Statistics.mean

Base.:+ — Method

Base.+(x::DisciplineLabor , y::DisciplineLabor)

Extends Base.+ to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add two DisciplineLabor objects together.

Base.:+ — Method

Base.+(x::TeamLabor , y::DisciplineLabor)

Extends Base.+ to add TeamLabor and DisciplineLabor together.

Arguments

x::TeamLabor: TeamLabor object to add

y::DisciplineLabor: DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add the two objects together.

ResourceManagement.ReadAvailHours — Method

ReadAvailHours(fName::String)

Function to read Available hours on a given month from SAP. Report should be saved in CSV format. YLFUTILDEPT is used to pull available hours from. Since SAP report only goes fwd 12 months, spreadhsheet can be edited to include avail hours beyond 12 months.

Arguments

fName::String: filename of the report

Returns

df::DataFrame: DataFrame of the report

Example:

    df = ReadAvailHours("C:\Users\james\Desktop\labor_tracker.csv")

ResourceManagement.ReadLaborTracker — Method

ReadLaborTracker(fName::String)

Function to read ACTUAL_PLANNED hours from SAP. Report should be saved in CSV format. Rows with missing names of resources are ignored.

Arguments

fName::String: filename of the report

Returns

df::DataFrame: DataFrame of the report

Example:

    df = ReadLaborTracker("C:\Users\james\Desktop\labor_tracker.csv")

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::TeamLabor: TeamLabor Object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._getAvailMonthHours — Method

_getAvailMonthHours(df::DataFrame,  m::Int)

Function to read Available hours for each month from DataFrame.

Arguments

df::DataFrame: DataFrame (output of ReadAvailHours)

Returns

v::Vector: Vector with available hours for each month. Pos [1] is current month.

Example:

    V = getAvailMonthHours("C:\Users\james\Desktop\UTILREPORT_NOV.csv")

ResourceManagement._getEmployeePlannedHours — Method

_getEmployeePlannedHours(df::DataFrame, name::String, m::Int)

Function to filter df by Employee name.

Arguments

df::DataFrame: dataframe of report (output of ReadLaborTracker)

name::String: name of employee (need to match sap name)

m::Int: number of months

Returns

v::Vector: Vector with the hours for each month

p:: Vector: Vector with employee's projects

Example:

    vh, pv = getEmployeePlannedHours(dflabor, "Doe John", 24)

ResourceManagement._pad_zerosEnd — Method

Function to pad vector from the end with zeroes. Parameters: v: vector to pad n: number of zeroes to pad

ResourceManagement.fetchAndWritePlannedHours! — Method

fetchAndWritePlannedHours!(df::DataFrame, name::String, m::Int, D::LaborVariable)

Function to fetch planned hours from ACTUAL_PLANNED SAP report for a given Employee name and number of months. Function will then save information to the LaborVariable object. LaborVariable.FwdHoursForecast[1] will hold the current month.

Arguments

df::DataFrame: df labor DataFrame

name::String: Name of Employee

m::Int: number of months to fetch and save

D::LaborVariable: LaborVariable object to save information to

Returns

vh: Array of vectors with Planned Hours for each month for projects

pv: Array of vectors with Projects

D::LaborVariable : LaborVariable object with the updated information.

Throws

undefined Projects: if projects do not match or are not found.

ResourceManagement.getCapacity — Method

getCapacity(D::LaborVariable)

Get Capacity of Resource from LaborVariable object. Capacity is defined as the Difference between the sum of Available Hours for a given month and the sum of Planned Hours for a given month.

Arguments

D::LaborVariable: LaborVariable object to get information from

Returns

v::Vector: Vector with capacity for each month

ResourceManagement.getFwdPlannedHours — Method

getFwdPlannedHours(D::LaborVariable, proj::String)

Function to get planned hours for a given project from a LaborVariable object.

Arguments

D::LaborVariable: LaborVariable object to get information from

proj::String: Project to get information for. If no project is given ("") or found, function will return all projects.

Returns

v::Vector: Vector with planned hours for the project

ResourceManagement.getUtilization — Method

getUtilization(D::LaborVariable, proj::String)

Function to get utilization for a given project from a LaborVariable object. If the given project is not found, function will return utilization for all projects.

Arguments

D::LaborVariable: LaborVariable object to get information from

proj::String: Project to get information for. If no project is given ("") or found, function will return all projects.

TODO: V::Vector: Vector with utilization for the project (TOO DEPENDANT ON OTHER FUNCTION)

Returns

v::Vector: Vector with utilization (percent) for the project

Example:

    V = getUtilization(JohnDoe, "")

ResourceManagement.writeAvailableFwdHours! — Method

writeAvailableFwdHours!(D::LaborVariable, df::DataFrame, m::Int)

Function will take an Array of Floats containing available monthly hours and and a LaborVariable and write the available hours to the LaborVariable FwdHoursAvailable field. This is a mutable function.

Arguments

D::LaborVariable: LaborVariable object to write to

df::DataFrame: DataFrame (output of ReadAvailHours)

m::Int: number of months to write to

Statistics.mean — Method

Statistics.mean(x::Float64, y::Float64)

Extends Statistics.mean to return the mean of two numbers.

Arguments

x::Float64: First number

y::Float64: Second number

Returns

Float64: Mean of two numbers