ResourceManagement

Documentation for ResourceManagement.

Base.:+

Base.:+

Base.:+

ResourceManagement.TeamDump

ResourceManagement._ProgramBuilder

ResourceManagement._TeamBuilder

ResourceManagement._TeamBuilder

ResourceManagement._getAvailMonthHours

ResourceManagement._getCostTrackerFromDf

ResourceManagement._getEmployeeHoursFromDf

ResourceManagement._pad_zerosEnd

ResourceManagement.getCapacity

ResourceManagement.getFwdPlannedHours

ResourceManagement.getFwdUtilization

ResourceManagement.getRevActualHours

ResourceManagement.getRevPlannedHours

ResourceManagement.writeAvailableFwdHours!

Statistics.mean

Base.:+ — Method

Base.+(x::DisciplineLabor , y::DisciplineLabor)

Extends Base.+ to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add two DisciplineLabor objects together.

Base.:+ — Method

Base.+(p1::Project , p2::Project)

Extends Base.+ to add two Project objects together.

Arguments

p1::Project: First Project object to add

p2::Project: Second Project object to add

Returns

Nothing. Function calls _ProgramBuilder to add two Project objects together.

Base.:+ — Method

Base.+(x::TeamLabor , y::DisciplineLabor)

Extends Base.+ to add TeamLabor and DisciplineLabor together.

Arguments

x::TeamLabor: TeamLabor object to add

y::DisciplineLabor: DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add the two objects together.

ResourceManagement.TeamDump — Method

function TeamDump(Team::TeamLabor, target::Symbol)

Function to iterate over all Team Resources and output target field as Dict pairs

Arguments

Team::TeamLabor: TeamLabor object target::Symbol: Symbol of field to dump targets: :TotalFwdPlannedHours (monthly FwdPlannedHours vector for all projects)

Returns

Dict: Dict of Team Resources with target field Keys are Employee names Values are Dict of target field

ResourceManagement._ProgramBuilder — Method

_ProgramBuilder(p1::Project, p2::Project)

Function to build a Program object from two Project objects. A Program is a collection of Projects.

Arguments

p1::Project: Project object to build Program from

p2::Project: Project object to build Program from

Returns

p::Program: Program object built from p1 and p2

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::TeamLabor: TeamLabor Object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._getAvailMonthHours — Method

_getAvailMonthHours(df::DataFrame,  m::Int)

Function to read Available hours for each month from DataFrame.

Arguments

df::DataFrame: DataFrame (output of ReadAvailHours)

Returns

v::Vector: Vector with available hours for each month. Pos [1] is current month.

Example:

    V = getAvailMonthHours("C:\Users\james\Desktop\UTILREPORT_NOV.csv")

ResourceManagement._getCostTrackerFromDf — Method

_getCostTrackerFromDf(df::DataFrame, Pnumber::Int, Dept::String)

Function to return cost data relative to a given project number and department.

Arguments

df::DataFrame: DataFrame (output of ReadCostTracker)

Pnumber::Int: Project number to get cost data for

Dept::String: Department to get cost data for

Returns

df:: DataFrame: DataFrame with cost data for the given project and department

ResourceManagement._getEmployeeHoursFromDf — Method

Function to transform df into a vector of vectors rows of this vector will represent each month (18 months as default for rev report) then each row will have a vector of hours for each project. Hours are summed for each project.

function _getEmployeeHoursFromDf2(df::DataFrame, name::String, m::Int, col::Symbol)

Parameters

df: DataFrame

name: String

m: Int

col: Symbol

Returns

v: Vector of vectors

ResourceManagement._pad_zerosEnd — Method

Function to pad vector from the end with zeroes. Parameters: v: vector to pad n: number of zeroes to pad

ResourceManagement.getCapacity — Method

getCapacity(D::LaborVariable)

Get Capacity of Resource from LaborVariable object. Capacity is defined as the Difference between the sum of Available Hours for a given month and the sum of Planned Hours for a given month.

Arguments

D::LaborVariable: LaborVariable object to get information from

Returns

v::Vector: Vector with capacity for each month

ResourceManagement.getFwdPlannedHours — Function

getFwdPlannedHours(D::LaborVariable, proj::String)

Function to get planned hours for a given project from a LaborVariable object.

Arguments

D::LaborVariable: LaborVariable object to get information from

proj::String: Project to get information for. If no project is given ("") or found, function will return all projects.

Returns

v::Vector: Vector with planned hours for the project

ResourceManagement.getFwdUtilization — Function

getUtilization(D::LaborVariable, proj::String)

Function to get utilization for a given project from a LaborVariable object. If the given project is not found, function will return utilization for all projects.

Arguments

D::LaborVariable: LaborVariable object to get information from

proj::String: Project to get information for. If no project is given ("") or found, function will return all projects.

TODO: V::Vector: Vector with utilization for the project (TOO DEPENDANT ON OTHER FUNCTION)

Returns

v::Vector: Vector with utilization (percent) for the project

Example:

    V = getUtilization(JohnDoe, "")

ResourceManagement.getRevActualHours — Function

getRevActualHours(D::LaborVariable, proj::String)

Function to get planned hours for a given project from a LaborVariable object.

Arguments

D::LaborVariable: LaborVariable object to get information from

proj::String: Project to get information for. If no project is given ("") or found, function will return all projects.

Returns

v::Vector: Vector with planned hours for the project

ResourceManagement.getRevPlannedHours — Function

getRevPlannedHours(D::LaborVariable, proj::String)

Function to get planned hours for a given project from a LaborVariable object.

Arguments

D::LaborVariable: LaborVariable object to get information from

proj::String: Project to get information for. If no project is given ("") or found, function will return all projects.

Returns

v::Vector: Vector with planned hours for the project

ResourceManagement.writeAvailableFwdHours! — Method

writeAvailableFwdHours!(D::LaborVariable, df::DataFrame, m::Int)

Function will take an Array of Floats containing available monthly hours and and a LaborVariable and write the available hours to the LaborVariable FwdHoursAvailable field. This is a mutable function.

Arguments

D::LaborVariable: LaborVariable object to write to

df::DataFrame: DataFrame (output of ReadAvailHours)

m::Int: number of months to write to

Statistics.mean — Method

Statistics.mean(x::Float64, y::Float64)

Extends Statistics.mean to return the mean of two numbers.

Arguments

x::Float64: First number

y::Float64: Second number

Returns

Float64: Mean of two numbers