ResourceManagement

Documentation for ResourceManagement.

ResourceManagement.DisciplineLabor

ResourceManagement.TeamLabor

Base.:+

Base.:+

ResourceManagement._TeamBuilder

ResourceManagement._TeamBuilder

Statistics.mean

ResourceManagement.DisciplineLabor — Type

DisciplineLabor

Fields

Rate::Float64: stores the rate for the labor variable

BudgetHours::Float64: stores budget for the resource in hours

BudgetDollars::Float64: stores budget for the resource in Dollars

TravelBudgetDollars::Float64: stores budget for Travel in Dollars

ActualHours::Float64: stores actual incurred hours for resource

ActualHours::Float64: stores actual incurred hours for resource

ActualIncurredCost::Float64: stores actual incurred Costs

FwdHoursAvailable::Array: stores monthly Hours Available from Present going forward

FwdHoursForecast::Array: stores Forecast monthly Hours from Present month going forward

RevHoursForecast::Array: stores Forecast monthly Hours from Present month going backwards (reverse)

RevHoursAvailable::Array: stores Forecast monthly Hours from Present month going backwards (reverse)

FwdCostsForecast::Array: stores forecasted monthly Labor cost from Present going forward

RevCostsForecast::Array: stores forecasted monthly Labor cost from Present going backwards (reverse)

RevActualCostHours::Array: stores monthly actual labor costs from Present going backwards (reverse)

Dept::String: stores Dept name

Name::String: stores name of the resource

Projects::Array: stores projects for the resource

ResourceManagement.TeamLabor — Type

TeamLabor()

Generates a TeamLabor object.

Arguments

Team::Array: Array of DisciplineLabor objects

Name::String: Name of the Team

Dept::String: Department of the Team

Rate::Float64: Engineering Labor Rate for the Team

BudgetHours::Float64: Labor Budget in LaborBudgetHours

BudgetDollars::Float64: Labor Budget in Dollars

TravelBudgetDollars::Float64: travel Budget in Dollars

ActualHours::Float64: Total Incurred Hours

ActualIncurredCost::Float64: Total Incurred Cost

Projects::Array : Array of Projects for the Team

FwdHoursAvailable::Array: stores Fwd Hours Available on a given month.

FwdHoursForecast::Array: stores Forecast monthly Hours from Present month going forward

RevHoursForecast::Array: stores Forecast monthly Hours from Present month going backwards (reverse)

RevHoursAvailable::Array: stores Available monthly Hours from Present month going backwards (reverse)

FwdCostsForecast::Array: stores Forecasted monthly Labor cost from Present going forward

RevCostsForecast::Array: stores Forecasted monthly Labor cost from Present going backwards (reverse)

RevActualHours::Array: stores Actual monthly Hours from Present month going backwards (reverse)

RevActualCostHours::Array: stores Actual monthly Labor cost from Present going backwards (reverse)

Returns

TeamLabor: TeamLabor object

Base.:+ — Method

Base.+(x::DisciplineLabor , y::DisciplineLabor)

Extends Base.+ to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add two DisciplineLabor objects together.

Base.:+ — Method

Base.+(x::TeamLabor , y::DisciplineLabor)

Extends Base.+ to add TeamLabor and DisciplineLabor together.

Arguments

x::TeamLabor: TeamLabor object to add

y::DisciplineLabor: DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add the two objects together.

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::TeamLabor: TeamLabor Object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

Statistics.mean — Method

Statistics.mean(x::Float64, y::Float64)

Extends Statistics.mean to return the mean of two numbers.

Arguments

x::Float64: First number

y::Float64: Second number

Returns

Float64: Mean of two numbers