ResourceManagement

Documentation for ResourceManagement.

ResourceManagement.DisciplineLabor

ResourceManagement.TeamLabor

Base.:+

Base.:+

ResourceManagement.ReadLaborTracker

ResourceManagement._TeamBuilder

ResourceManagement._TeamBuilder

ResourceManagement._getEmployeePlannedHours

ResourceManagement._pad_zerosEnd

ResourceManagement.fetchAndWritePlannedHours!

Statistics.mean

ResourceManagement.DisciplineLabor — Type

DisciplineLabor

Fields

Rate::Float64: stores the rate for the labor variable

BudgetHours::Float64: stores budget for the resource in hours

BudgetDollars::Float64: stores budget for the resource in Dollars

TravelBudgetDollars::Float64: stores budget for Travel in Dollars

ActualHours::Float64: stores actual incurred hours for resource

ActualHours::Float64: stores actual incurred hours for resource

ActualIncurredCost::Float64: stores actual incurred Costs

FwdHoursAvailable::Array: stores monthly Hours Available from Present going forward

FwdHoursForecast::Array: stores Forecast monthly Hours from Present month going forward

RevHoursForecast::Array: stores Forecast monthly Hours from Present month going backwards (reverse)

RevHoursAvailable::Array: stores Forecast monthly Hours from Present month going backwards (reverse)

FwdCostsForecast::Array: stores forecasted monthly Labor cost from Present going forward

RevCostsForecast::Array: stores forecasted monthly Labor cost from Present going backwards (reverse)

RevActualCostHours::Array: stores monthly actual labor costs from Present going backwards (reverse)

Dept::String: stores Dept name

Name::String: stores name of the resource

Projects::Array: stores projects for the resource

ResourceManagement.TeamLabor — Type

TeamLabor()

Generates a TeamLabor object.

Arguments

Team::Array: Array of DisciplineLabor objects

Name::String: Name of the Team

Dept::String: Department of the Team

Rate::Float64: Engineering Labor Rate for the Team

BudgetHours::Float64: Labor Budget in LaborBudgetHours

BudgetDollars::Float64: Labor Budget in Dollars

TravelBudgetDollars::Float64: travel Budget in Dollars

ActualHours::Float64: Total Incurred Hours

ActualIncurredCost::Float64: Total Incurred Cost

Projects::Array : Array of Projects for the Team

FwdHoursAvailable::Array: stores Fwd Hours Available on a given month.

FwdHoursForecast::Array: stores Forecast monthly Hours from Present month going forward

RevHoursForecast::Array: stores Forecast monthly Hours from Present month going backwards (reverse)

RevHoursAvailable::Array: stores Available monthly Hours from Present month going backwards (reverse)

FwdCostsForecast::Array: stores Forecasted monthly Labor cost from Present going forward

RevCostsForecast::Array: stores Forecasted monthly Labor cost from Present going backwards (reverse)

RevActualHours::Array: stores Actual monthly Hours from Present month going backwards (reverse)

RevActualCostHours::Array: stores Actual monthly Labor cost from Present going backwards (reverse)

Returns

TeamLabor: TeamLabor object

Base.:+ — Method

Base.+(x::DisciplineLabor , y::DisciplineLabor)

Extends Base.+ to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add two DisciplineLabor objects together.

Base.:+ — Method

Base.+(x::TeamLabor , y::DisciplineLabor)

Extends Base.+ to add TeamLabor and DisciplineLabor together.

Arguments

x::TeamLabor: TeamLabor object to add

y::DisciplineLabor: DisciplineLabor object to add

Returns

Nothing. Function calls _TeamBuilder to add the two objects together.

ResourceManagement.ReadLaborTracker — Method

ReadLaborTracker(fName::String)

Function to read ACTUAL_PLANNED hours from SAP. Report should be saved in CSV format. Rows with missing names of resources are ignored.

Arguments

fName::String: filename of the report

Returns

df::DataFrame: DataFrame of the report

Example:

    df = ReadLaborTracker("C:\Users\james\Desktop\labor_tracker.csv")

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::DisciplineLabor: First DisciplineLabor object

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._TeamBuilder — Method

_TeamBuilder(x::DisciplineLabor, y::DisciplineLabor)

Function to add two DisciplineLabor objects together.

Arguments

x::TeamLabor: TeamLabor Object to add

y::DisciplineLabor: Second DisciplineLabor object to add

Returns

T::TeamLabor: Team Labor object

ResourceManagement._getEmployeePlannedHours — Method

_getEmployeePlannedHours(df::DataFrame, name::String, m::Int)

Function to filter df by Employee name.

Arguments

df::DataFrame: dataframe of report (output of ReadLaborTracker)

name::String: name of employee (need to match sap name)

m::Int: number of months

Returns

v::Vector: Vector with the hours for each month

p:: Vector: Vector with employee's projects

Example:

    vh, pv = getEmployeePlannedHours(dflabor, "Doe John", 24)

ResourceManagement._pad_zerosEnd — Method

Function to pad vector from the end with zeroes. Parameters: v: vector to pad n: number of zeroes to pad

ResourceManagement.fetchAndWritePlannedHours! — Method

fetchAndWritePlannedHours!(df::DataFrame, name::String, m::Int, D::LaborVariable)

Function to fetch planned hours from ACTUAL_PLANNED SAP report for a given Employee name and number of months. Function will then save information to the LaborVariable object. LaborVariable.FwdHoursForecast[1] will hold the current month.

Arguments

df::DataFrame: df labor DataFrame

name::String: Name of Employee

m::Int: number of months to fetch and save

D::LaborVariable: LaborVariable object to save information to

Returns

vh: Array of vectors with Planned Hours for each month for projects

pv: Array of vectors with Projects

D::LaborVariable : LaborVariable object with the updated information.

Throws

undefined Projects: if projects do not match or are not found.

Statistics.mean — Method

Statistics.mean(x::Float64, y::Float64)

Extends Statistics.mean to return the mean of two numbers.

Arguments

x::Float64: First number

y::Float64: Second number

Returns

Float64: Mean of two numbers